Remove test_timedelta.py
